Add forking startup wrapper for Gimp
When Gimp is is launched from webmenu, it stays as webmenu's child and
freezes (tool windows do not load properly). This seems to happend when
the current locale is fi_FI, en_GB works well.

However, the freeze disappears when webmenu is killed and init adopts
gimp. Moreover, the current version of node webkit has buggy spawning
code, it does not detach spawned process from its parent as expected.